# api-spec/dist/openapi.yaml
openapi: 3.0.0
info:
  title: Hexlet Basics API
  description: |-
    Hexlet Basics API.

    Contract-first (ADR-0001): this TypeSpec compiles to `dist/openapi.yaml`
    (`pnpm spec`), which drives BOTH the Go server (ogen validators) and the TS
    frontend client + zod (hey-api, `pnpm generate`). Never hand-edit `dist/`.

    Naming: the domain concept is `Course`. The legacy Rails app called it
    `Language`; we keep `Language` ONLY in URL paths (e.g. `/languages`) for
    backward compatibility (ADR-0002), but every type/operation/schema uses
    `Course`. The full model/enum set lives in `models.tsp`.
  version: 0.0.0
...
paths:
...
  /admin/attachments:
    post:
      operationId: adminUploadAttachment
      description: Upload a file; returns the attachment to reference from a form.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '401':
          description: The request is not authenticated (no/invalid session cookie).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The caller is authenticated but lacks permission for this resource.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: |-
            Field-level validation errors, keyed by field name (each value is the list
            of messages for that field). Returned when a write fails validation —
            including constraints the schema cannot express, like uniqueness.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        default:
          description: |-
            Default error response shared by every operation.

            `@error` emits an OpenAPI `default` response, keeping central transport
            failures typed without enumerating every status on every operation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - 'Admin: Attachments'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttachmentUploadForm'
      security:
        - AdminSession: []
          XsrfToken: []
...
components:
...
  schemas:
...
    Attachment:
      type: object
      required:
        - id
        - url
        - filename
        - contentType
        - byteSize
      properties:
        id:
          type: integer
          format: int32
        url:
          type: string
        filename:
          type: string
        contentType:
          type: string
        byteSize:
          type: integer
          format: int64
      description: A stored blob (cover image, outcomes image) returned by the uploader.
    AttachmentUploadForm:
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: URI identifying the problem type. `about:blank` means the HTTP status.
        title:
          type: string
          description: Short, localized summary for this problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code generated for this occurrence.
        detail:
          type: string
          description: Optional occurrence-specific explanation that is safe for clients.
        instance:
          type: string
          description: Optional URI identifying this particular occurrence.
      description: |-
        RFC 9457 problem details returned by the transport when request decoding,
        authorization, persistence, or an unexpected server failure prevents an
        operation from producing its declared domain response.
...
    ValidationError:
      type: object
      required:
        - errors
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      description: |-
        Field-level validation errors, keyed by field name (each value is the list
        of messages for that field). Returned when a write fails validation —
        including constraints the schema cannot express, like uniqueness.
  securitySchemes:
    UserSession:
      type: apiKey
      in: cookie
      name: JWT
    XsrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    AdminSession:
      type: apiKey
      in: cookie
      name: JWT